containers/guardrails/requirements.txt
fastapi
uvicorn
nemoguardrails
langchain-openai
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
openinference-instrumentation-langchain
openinference-instrumentation-openai
opentelemetry-instrumentation-httpx
